chore: fix tests duplication
